+++ src/main/java/org/grupofort/adapters/controllers/RegisterPaymentRequest.java
@@ -1,13 +1,11 @@
 package org.grupofort.adapters.controllers;
 
-import java.math.BigDecimal;
-
 public record RegisterPaymentRequest(
         int dia,
         int mes ,
         int ano,
         long codass,
-        BigDecimal valorPago
+        double valorPago
 )
 {
